BotFDS2.0/bot/requirements.txt
aiogram==3.0.0
python-dotenv==0.19.0
sqlalchemy[asyncio]==2.0.0
alembic==1.8.1
apscheduler==3.9.0